+++ src/datumaro/plugins/data_formats/datumaro/exporter.py
@@ -20,6 +20,7 @@ from datumaro.components.annotation import (
     Annotation,
     Bbox,
     Caption,
+    Cuboid2D,
     Cuboid3d,
     Ellipse,
     HashKey,
@@ -37,6 +38,7 @@ from datumaro.components.annotation import (
 from datumaro.components.crypter import NULL_CRYPTER
 from datumaro.components.dataset_base import DatasetItem
 from datumaro.components.dataset_item_storage import ItemStatus
+from datumaro.components.errors import PathSeparatorInSubsetNameError
 from datumaro.components.exporter import ExportContextComponent, Exporter
 from datumaro.components.media import Image, MediaElement, PointCloud, Video, VideoFrame
 from datumaro.util import cast, dump_json_file
@@ -184,7 +186,8 @@ class _SubsetWriter:
 
             if context.save_media:
                 fname = context.make_video_filename(item)
-                context.save_video(item, fname=fname, subdir=item.subset)
+                subdir = item.subset.replace(os.sep, "_") if item.subset else None
+                context.save_video(item, fname=fname, subdir=subdir)
                 item.media = Video(
                     path=fname,
                     step=video._step,
@@ -199,7 +202,8 @@ class _SubsetWriter:
 
             if context.save_media:
                 fname = context.make_video_filename(item)
-                context.save_video(item, fname=fname, subdir=item.subset)
+                subdir = item.subset.replace(os.sep, "_") if item.subset else None
+                context.save_video(item, fname=fname, subdir=subdir)
                 item.media = VideoFrame(Video(fname), video_frame.index)
 
             yield
@@ -209,8 +213,9 @@ class _SubsetWriter:
 
             if context.save_media:
                 # Temporarily update image path and save it.
-                fname = context.make_image_filename(item)
-                context.save_image(item, encryption=encryption, fname=fname, subdir=item.subset)
+                fname = context.make_image_filename(item, name=str(item.id).replace(os.sep, "_"))
+                subdir = item.subset.replace(os.sep, "_") if item.subset else None
+                context.save_image(item, encryption=encryption, fname=fname, subdir=subdir)
                 item.media = Image.from_file(path=fname, size=image._size)
 
             yield
@@ -219,14 +224,18 @@ class _SubsetWriter:
             pcd = item.media_as(PointCloud)
 
             if context.save_media:
-                pcd_fname = context.make_pcd_filename(item)
-                context.save_point_cloud(item, fname=pcd_fname, subdir=item.subset)
+                pcd_name = str(item.id).replace(os.sep, "_")
+                pcd_fname = context.make_pcd_filename(item, name=pcd_name)
+                subdir = item.subset.replace(os.sep, "_") if item.subset else None
+                context.save_point_cloud(item, fname=pcd_fname, subdir=subdir)
 
                 extra_images = []
                 for i, extra_image in enumerate(pcd.extra_images):
                     extra_images.append(
                         Image.from_file(
-                            path=context.make_pcd_extra_image_filename(item, i, extra_image)
+                            path=context.make_pcd_extra_image_filename(
+                                item, i, extra_image, name=f"{pcd_name}/extra_image_{i}"
+                            )
                         )
                     )
 
@@ -311,6 +320,8 @@ class _SubsetWriter:
                 converted_ann = self._convert_ellipse_object(ann)
             elif isinstance(ann, HashKey):
                 continue
+            elif isinstance(ann, Cuboid2D):
+                converted_ann = self._convert_cuboid_2d_object(ann)
             else:
                 raise NotImplementedError()
             annotations.append(converted_ann)
@@ -435,6 +446,18 @@ class _SubsetWriter:
     def _convert_ellipse_object(self, obj: Ellipse):
         return self._convert_shape_object(obj)
 
+    def _convert_cuboid_2d_object(self, obj: Cuboid2D):
+        converted = self._convert_annotation(obj)
+
+        converted.update(
+            {
+                "label_id": cast(obj.label, int),
+                "points": obj.points,
+                "z_order": obj.z_order,
+            }
+        )
+        return converted
+
 
 class _StreamSubsetWriter(_SubsetWriter):
     def __init__(
@@ -492,18 +515,27 @@ class DatumaroExporter(Exporter):
             default_image_ext=self._default_image_ext,
         )
 
+        if os.path.sep in subset:
+            raise PathSeparatorInSubsetNameError(subset)
+
         return (
             _SubsetWriter(
                 context=self,
                 subset=subset,
-                ann_file=osp.join(self._annotations_dir, subset + self.PATH_CLS.ANNOTATION_EXT),
+                ann_file=osp.join(
+                    self._annotations_dir,
+                    subset + self.PATH_CLS.ANNOTATION_EXT,
+                ),
                 export_context=export_context,
             )
             if not self._stream
             else _StreamSubsetWriter(
                 context=self,
                 subset=subset,
-                ann_file=osp.join(self._annotations_dir, subset + self.PATH_CLS.ANNOTATION_EXT),
+                ann_file=osp.join(
+                    self._annotations_dir,
+                    subset + self.PATH_CLS.ANNOTATION_EXT,
+                ),
                 export_context=export_context,
             )
         )