+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## \[Q4 2024 Release 1.9.1\]
-### New features
+## \[Unreleased\]
 
+### New features
+- Support KITTI 3D format
+  (<https://github.com/openvinotoolkit/datumaro/pull/1619>)
+- Add PseudoLabeling transform for unlabeled dataset
+  (<https://github.com/openvinotoolkit/datumaro/pull/1594>)
+
+### Enhancements
+- Raise an appropriate error when exporting a datumaro dataset if its subset name contains path separators.
+  (<https://github.com/openvinotoolkit/datumaro/pull/1615>)
+- Update docs for transform plugins
+  (<https://github.com/openvinotoolkit/datumaro/pull/1599>)
+
+### Bug fixes
+
+## Q4 2024 Release 1.9.1
 ### Enhancements
 - Support multiple labels for kaggle format
   (<https://github.com/openvinotoolkit/datumaro/pull/1607>)
@@ -22,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### New features
 - Add a new CLI command: datum format
   (<https://github.com/openvinotoolkit/datumaro/pull/1570>)
+- Add a new Cuboid2D annotation type
+  (<https://github.com/openvinotoolkit/datumaro/pull/1601>)
 - Support language dataset for DmTorchDataset
   (<https://github.com/openvinotoolkit/datumaro/pull/1592>)